KMM-98: - Added sample videos list and detailed screen - Refactored navigation graph structure - Added Button in launch details screen to watch launch video - Refactored packages structure - Removed implementing sharedfirebase kmm module - Bumped dependencies
