+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/core/constants/ComposeConstants.kt
@@ -1,0 +1,5 @@
+package ru.lyrian.kotlinmultiplatformsandbox.android.core.constants
+
+object ComposeConstants {
+    const val SMALL_BUTTON_CORNER_PERCENT = 50
+}
