--- androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/core/ui/navigation/bottom_navigation/BottomNavTitles.kt
@@ -1,7 +1,0 @@
-package ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.bottom_navigation
-
-object BottomNavTitles {
-    const val LIST_TITLE = "Launches"
-    const val FAVORITES_TITLE = "Favorites"
-    const val PROFILE_TITLE = "Profile"
-}
