--- androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/core/ui/navigation/destinations/NavDestinationsArgs.kt
@@ -1,7 +1,0 @@
-package ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.destinations
-
-object NavDestinationsArgs {
-    object Details {
-        const val LAUNCH_ID_ARG = "launchId"
-    }
-}
