+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/core/ui/navigation/nav_graphs/DetailsNavGraph.kt
@@ -1,31 +1,29 @@
 package ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.nav_graphs
 
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.destinations.NavDestinations
-import ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.destinations.NavDestinationsArgs.Details.LAUNCH_ID_ARG
-import ru.lyrian.kotlinmultiplatformsandbox.android.feature.launch_details.presentation.ui.LaunchDetailsScreen
+import ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.destinations.Destinations.RootGraph.HomeGraph.LaunchesDetailsGraph
+import ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.screens.LocalScaffoldPaddings
+import ru.lyrian.kotlinmultiplatformsandbox.android.feature.launches.details.presentation.ui.LaunchDetailsScreen
 
-fun NavGraphBuilder.detailsNavGraph(
-    onNavigateBack: () -> Unit
+fun NavGraphBuilder.launchDetailsNavGraph(
+    onNavigateBack: () -> Unit,
+    onWatchVideoClick: (String) -> Unit
 ) {
     navigation(
-        route = NavGraphsDestinations.DETAILS,
-        startDestination = NavDestinations.DetailsNavGraph.DETAILS
+        route = LaunchesDetailsGraph.GRAPH_ROUTE,
+        startDestination = LaunchesDetailsGraph.DETAILS
     ) {
-        composable("${NavDestinations.DetailsNavGraph.DETAILS}/{$LAUNCH_ID_ARG}") {
+        composable("${LaunchesDetailsGraph.DETAILS}/{${LaunchesDetailsGraph.DetailsArgs.LAUNCH_ID}}") {
             LaunchDetailsScreen(
                 onNavigateBackClick = onNavigateBack,
-                modifier = Modifier.fillMaxSize()
+                onWatchVideoClick = onWatchVideoClick,
+                modifier = Modifier.fillMaxSize().padding(LocalScaffoldPaddings.current),
             )
         }
     }
-}
-
-fun NavController.navigateToLaunchDetails(launchId: String) {
-    this.navigate("${NavDestinations.DetailsNavGraph.DETAILS}/$launchId")
 }
