--- androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/core/ui/navigation/nav_graphs/NavGraphsDestinations.kt
@@ -1,7 +1,0 @@
-package ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.nav_graphs
-
-object NavGraphsDestinations {
-    const val ROOT = "rootGraph"
-    const val HOME = "homeGraph"
-    const val DETAILS = "detailsGraph"
-}
