+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/core/ui/screens/HomeScreen.kt
@@ -1,7 +1,8 @@
 package ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.screens
 
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.material.SnackbarHostState
 import androidx.compose.material.rememberScaffoldState
@@ -15,25 +16,28 @@ import ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.bottom_na
 import ru.lyrian.kotlinmultiplatformsandbox.android.core.ui.navigation.nav_graphs.HomeNavGraph
 
 val LocalSnackBarState = compositionLocalOf<SnackbarHostState> { error("No SnackbarHostState provided") }
+val LocalScaffoldPaddings = compositionLocalOf<PaddingValues> { error("No PaddingValues provided") }
 
+@SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
 fun HomeScreen(navController: NavHostController = rememberNavController()) {
     val scaffoldState = rememberScaffoldState()
 
-    CompositionLocalProvider(LocalSnackBarState provides scaffoldState.snackbarHostState) {
-        Scaffold(
-            scaffoldState = scaffoldState,
-            bottomBar = {
-                BottomNavigationBar(navController = navController)
-            }
+    Scaffold(
+        scaffoldState = scaffoldState,
+        bottomBar = {
+            BottomNavigationBar(navController = navController)
+        }
+    ) {
+        CompositionLocalProvider(
+            LocalSnackBarState provides scaffoldState.snackbarHostState,
+            LocalScaffoldPaddings provides it
         ) {
             HomeNavGraph(
                 navController = navController,
                 modifier = Modifier
-                    .padding(it)
                     .fillMaxSize()
             )
         }
     }
-
 }
