+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/di/AndroidModules.kt
@@ -1,11 +1,11 @@
 package ru.lyrian.kotlinmultiplatformsandbox.android.di
 
-import ru.lyrian.kotlinmultiplatformsandbox.android.feature.launch_details.di.launchDetailsModule
-import ru.lyrian.kotlinmultiplatformsandbox.android.feature.launches_list.di.launchesListModule
+import ru.lyrian.kotlinmultiplatformsandbox.android.feature.launches.di.launchesModule
 import ru.lyrian.kotlinmultiplatformsandbox.android.feature.profile.di.profileModule
+import ru.lyrian.kotlinmultiplatformsandbox.android.feature.videos.di.videosModule
 
 fun androidModules() = listOf(
-    launchesListModule,
-    launchDetailsModule,
-    profileModule
+    launchesModule,
+    profileModule,
+    videosModule
 )
