--- androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/feature/launch_details/presentation/model/LaunchDetailsEvent.kt
@@ -1,5 +1,0 @@
-package ru.lyrian.kotlinmultiplatformsandbox.android.feature.launch_details.presentation.model
-
-sealed interface LaunchDetailsEvent {
-    data class ShowErrorMessage(val message: String) : LaunchDetailsEvent
-}
