+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/feature/launches/details/presentation/model/LaunchDetailsEvent.kt
@@ -1,0 +1,5 @@
+package ru.lyrian.kotlinmultiplatformsandbox.android.feature.launches.details.presentation.model
+
+sealed interface LaunchDetailsEvent {
+    data class ShowErrorMessage(val message: String) : LaunchDetailsEvent
+}
