+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/feature/launches/list/presentation/ui/LaunchesListEvent.kt
@@ -1,0 +1,5 @@
+package ru.lyrian.kotlinmultiplatformsandbox.android.feature.launches.list.presentation.ui
+
+sealed interface LaunchesListEvent {
+    data class ShowToast(val message: String): LaunchesListEvent
+}
