--- androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/feature/launches_list/presentation/ui/LaunchesListEvent.kt
@@ -1,5 +1,0 @@
-package ru.lyrian.kotlinmultiplatformsandbox.android.feature.launches_list.presentation.ui
-
-sealed interface LaunchesListEvent {
-    data class ShowToast(val message: String): LaunchesListEvent
-}
