+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/feature/profile/presentation/viewmodel/ProfileViewModel.kt
@@ -15,10 +15,7 @@ import ru.lyrian.kotlinmultiplatformsandbox.feature.profile.domain.Profile
 import ru.lyrian.kotlinmultiplatformsandbox.feature.profile.domain.ProfileInteractor
 
 class ProfileViewModel(
-    private val profileInteractor: ProfileInteractor,
-
-    // Commented example of crashlytics non fatal error. Use in release build variant
-    // private val firebaseIntegrationInteractor: FirebaseIntegrationInteractor
+    private val profileInteractor: ProfileInteractor
 ) : ViewModel() {
 
     private var savedProfile: Profile? = null
@@ -58,8 +55,6 @@ class ProfileViewModel(
                     )
                 }
         }
-
-
     }
 
     fun updateUserName(newUserName: String) {
@@ -73,14 +68,7 @@ class ProfileViewModel(
     }
 
     fun saveProfile() {
-        // Commented example of crashlytics non fatal error. Use in release build variant
-        // firebaseIntegrationInteractor.firebaseCrashlyticsTest()
-
         viewModelScope.launch {
-            // Commented example of Firebase usage from KMM module
-//            FirestoreClient().addUser("petya")
-//            FirestoreClient().getFirstUser()
-
             profile?.let { profileInteractor.saveProfile(it) }
             profileInteractor
                 .getProfile()
