+++ androidApp/src/main/java/ru/lyrian/kotlinmultiplatformsandbox/android/feature/videos/di/VideosModule.kt
@@ -1,0 +1,7 @@
+package ru.lyrian.kotlinmultiplatformsandbox.android.feature.videos.di
+
+import org.koin.dsl.module
+
+val videosModule = module {
+
+}
